refactor: rename internal templating package to dashboard
