+++ release-automation/internal/dashboard/tractusx.go
@@ -17,7 +17,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package templating
+package dashboard
 
 import (
 	"context"
