+++ release-automation/internal/dashboard/types.go
@@ -17,7 +17,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package templating
+package dashboard
 
 import (
 	"github.com/eclipse-tractusx/tractusx-quality-checks/pkg/tractusx"
